+++ CHANGELOG.md
@@ -4,7 +4,13 @@
 
 ### Fixed
 
-- [TD-1749] Added business concept current version id to search response for relations with business_concept as target
+- [TD-1749] Added business concept current version id to search response for relations with business_concept as source/target
+
+## [2.20.0] 2019-05-27
+
+### Added
+
+- [TD-1535] Check ingest permissions
 
 ## [2.19.0] 2019-05-14